moving linq extention to common namespace
